Add the ability to create and auth with opaque tokens
DIfferent than JWT in that:
- A user can have many of these
- They can individually be activated/inactivated
- They can be individually deleted
- They can be created with custom expiration times (via UI)
- They don't encode any information. All information lives in the database
- The token is modeled as:
`<locator (uuid)>.<key (uuid)>`
As a key is validated, it extracts the locator and uses
that to lookup the key in the DB.  That entry contains an
Argon2 hash.  The key is then compared to the hash. (it
also ensures that the token is active, not expired, etc)