Tracking user weekly cells-processed quotas (#155)
Users now have an associated cell quota, which is a count of cells they can process during a week.  Their quota resets at midnight on the weekday that matches the day their account was created.  If an annotate request is received that would cause the user to go over their quota for the week, a 403 response is returned with a message that includes their remaining quota for the week.  To account for cells currently being processed, the user activity table now also tracks an event for each row, either STARTED, SUCCEEDED, or FAILED.  STARTED events are matched to SUCCEEDED/FAILED events by a request ID.  The rows are matched up when calculating the user's remaining quota.  Only SUCCEEDED and unmatched STARTED rows are counted.  There is also a new endpoint for getting a user's total and remaining quota, along with their quota reset date.

---------

Co-authored-by: Nick Malfroy-Camine <nmalfroy@broadinstitute.org>